Fixed some warnings on Windows builds.

+++ UGen/basics/ugen_InlineUnaryOps.h
@@ -186,7 +186,7 @@ inline float ampdb(const float amp) throw()			{	return (float)::log10(amp) * 20.
 inline float dbamp(const float db) throw()			{	return (float)::pow(10.f, db * .05f);						}
 inline double ampdb(const double amp) throw()		{	return ::log10(amp) * 20.;									}
 inline double dbamp(const double db) throw()		{	return ::pow(10, db * .05);									}
-inline double ampdb(const int amp) throw()			{	return ::log10(amp) * 20.;									}
+inline double ampdb(const int amp) throw()			{	return ::log10((double)amp) * 20.;									}
 inline double dbamp(const int db) throw()			{	return ::pow(10, db * .05);									}
 
 inline float deg2rad(const float deg) throw()		{	return deg * (float)oneOver360 * (float)twoPi;				}
